Refactorting - Part 2
